automoji: fix max limiting
We were setting max by passing it to es.all *and* here.  This meant that
if we requested a max of 1 but *also* had a required emoji (like we
often do with 🎰) the required emoji ends up taking up a slot, and in
fact ends up not required.  Now required emoji effectively do not end up
counting against our max.